// frontend/src/api/model/adminPersonDTO.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Movie Review System API
 * Backend API for Movie Review Website
 * OpenAPI spec version: v1.0
 */

/**
 * 后台新增/更新影人请求参数
 */
export interface AdminPersonDTO {
  /**
   * 中文名
   * @minLength 0
   * @maxLength 100
   */
  name: string;
  /**
   * 性别
   * @minLength 0
   * @maxLength 20
   */
  sex?: string;
  /**
   * 英文名
   * @minLength 0
   * @maxLength 100
   */
  nameEn?: string;
  /**
   * 中文名(备用)
   * @minLength 0
   * @maxLength 100
   */
  nameZh?: string;
  /**
   * 出生日期
   * @minLength 0
   * @maxLength 50
   */
  birth?: string;
  /**
   * 出生地
   * @minLength 0
   * @maxLength 200
   */
  birthplace?: string;
  /**
   * 职业
   * @minLength 0
   * @maxLength 200
   */
  profession?: string;
  /**
   * 影人简介
   * @minLength 0
   * @maxLength 20000
   */
  biography?: string;
  /**
   * 影人头像 URL
   * @minLength 0
   * @maxLength 500
   */
  avatar?: string;
}